fix(pci-load-balancer): listener section pair testing returns
DTCORE-2777
