+++ packages/manager/apps/pci-load-balancer/src/pages/detail/listeners/delete/DeleteListener.page.tsx
@@ -10,12 +10,16 @@ import { useDeleteListener, useListener } from '@/api/hook/useListener';
 
 export default function DeleteListenerPage() {
   const { addSuccess, addError } = useNotifications();
-  const { t: tListenerDelete } = useTranslation('listener-delete');
+  const { t } = useTranslation('listeners/delete');
+
   const navigate = useNavigate();
+
   const { projectId, loadBalancerId, listenerId, region } = useParams();
+
   const onClose = () => {
     navigate('..');
   };
+
   const { data: listener, isPending: isPendingListener } = useListener({
     projectId,
     region,
@@ -31,12 +35,20 @@ export default function DeleteListenerPage() {
     onError(error: ApiError) {
       addError(
         <Translation ns="load-balancer">
-          {(_t) =>
-            _t('octavia_load_balancer_global_error', {
-              message: error?.response?.data?.message || error?.message || null,
-              requestId: error?.config?.headers['X-OVH-MANAGER-REQUEST-ID'],
-            })
-          }
+          {(_t) => (
+            <span
+              dangerouslySetInnerHTML={{
+                __html: _t('octavia_load_balancer_global_error', {
+                  message: ((error.response as unknown) as {
+                    data: { message: string };
+                  }).data.message,
+                  requestId: ((error.response as unknown) as {
+                    headers: Record<string, unknown>;
+                  }).headers['x-ovh-queryid'],
+                }),
+              }}
+            />
+          )}
         </Translation>,
         true,
       );
@@ -69,25 +81,21 @@ export default function DeleteListenerPage() {
 
   return (
     <DeletionModal
-      title={tListenerDelete('octavia_load_balancer_listener_delete_title')}
+      title={t('octavia_load_balancer_listener_delete_title')}
       onConfirm={onConfirm}
       onClose={onClose}
       onCancel={onCancel}
       isPending={isPending}
       type="warning"
-      submitText={tListenerDelete(
-        'octavia_load_balancer_listener_delete_confirm',
-      )}
-      cancelText={tListenerDelete(
-        'octavia_load_balancer_listener_delete_cancel',
-      )}
+      submitText={t('octavia_load_balancer_listener_delete_confirm')}
+      cancelText={t('octavia_load_balancer_listener_delete_cancel')}
     >
       <OsdsText
         level={ODS_TEXT_LEVEL.body}
         color={ODS_THEME_COLOR_INTENT.text}
         size={ODS_TEXT_SIZE._400}
       >
-        {tListenerDelete('octavia_load_balancer_listener_delete_description', {
+        {t('octavia_load_balancer_listener_delete_description', {
           listener: listener?.name,
         })}
       </OsdsText>
